wip - rewriting main, get now playing track id
